Move draw functions from utils to visualize file ( like where it should belong )
