+++ Yolo/utils.cpp
@@ -1,6 +1,7 @@
 #include "utils.hpp"
 #include "object.hpp"
 #include "detector.hpp"
+#include "visualize.hpp"
 #include <iostream>
 #include <vector>
 #include <filesystem>
@@ -13,90 +14,6 @@
 #include <opencv2/highgui/highgui.hpp>
 
 #include <yaml-cpp/yaml.h>
-
-const unsigned char colors[81][3] = {
-    {56,  0,   255},
-    {226, 255, 0},
-    {0,   94,  255},
-    {0,   37,  255},
-    {0,   255, 94},
-    {255, 226, 0},
-    {0,   18,  255},
-    {255, 151, 0},
-    {170, 0,   255},
-    {0,   255, 56},
-    {255, 0,   75},
-    {0,   75,  255},
-    {0,   255, 169},
-    {255, 0,   207},
-    {75,  255, 0},
-    {207, 0,   255},
-    {37,  0,   255},
-    {0,   207, 255},
-    {94,  0,   255},
-    {0,   255, 113},
-    {255, 18,  0},
-    {255, 0,   56},
-    {18,  0,   255},
-    {0,   255, 226},
-    {170, 255, 0},
-    {255, 0,   245},
-    {151, 255, 0},
-    {132, 255, 0},
-    {75,  0,   255},
-    {151, 0,   255},
-    {0,   151, 255},
-    {132, 0,   255},
-    {0,   255, 245},
-    {255, 132, 0},
-    {226, 0,   255},
-    {255, 37,  0},
-    {207, 255, 0},
-    {0,   255, 207},
-    {94,  255, 0},
-    {0,   226, 255},
-    {56,  255, 0},
-    {255, 94,  0},
-    {255, 113, 0},
-    {0,   132, 255},
-    {255, 0,   132},
-    {255, 170, 0},
-    {255, 0,   188},
-    {113, 255, 0},
-    {245, 0,   255},
-    {113, 0,   255},
-    {255, 188, 0},
-    {0,   113, 255},
-    {255, 0,   0},
-    {0,   56,  255},
-    {255, 0,   113},
-    {0,   255, 188},
-    {255, 0,   94},
-    {255, 0,   18},
-    {18,  255, 0},
-    {0,   255, 132},
-    {0,   188, 255},
-    {0,   245, 255},
-    {0,   169, 255},
-    {37,  255, 0},
-    {255, 0,   151},
-    {188, 0,   255},
-    {0,   255, 37},
-    {0,   255, 0},
-    {255, 0,   170},
-    {255, 0,   37},
-    {255, 75,  0},
-    {0,   0,   255},
-    {255, 207, 0},
-    {255, 0,   226},
-    {255, 245, 0},
-    {188, 255, 0},
-    {0,   255, 18},
-    {0,   255, 75},
-    {0,   255, 151},
-    {255, 56,  0},
-    {245, 255, 0}
-};
 
 std::vector<std::string> IMG_FORMATS {"bmp", "dng", "jpg", "jpeg", "mpo", "png", "tif", "tiff", "webp", "pfm"};
 std::vector<std::string> VID_FORMATS {"asf", "avi", "gif", "m4v", "mkv", "mov", "mp4", "mpeg", "mpg", "ts", "wmv"};
@@ -148,65 +65,6 @@ int Utils::run() {
 //     return 0;
 // }
 
-void Utils::draw_objects(cv::Mat& bgr, const std::vector<Object>& objects, int colorMode) {
-    size_t objCount = objects.size();
-    LOG("Objects count = " << objCount << std::endl);
-
-    int color_index = 0;
-    for (size_t i = 0; i < objCount; i++) {
-        const Object& obj = objects[i];
-
-        char line[256];
-        //class-index confident center-x center-y box-width box-height
-        sprintf(line, "%i %f %i %i %i %i", obj.label, obj.prob, (int) round(obj.rect.tl().x), (int) round(obj.rect.tl().y), (int) round(obj.rect.br().x), (int) round(obj.rect.br().y));
-
-        LOG(line << std::endl);
-
-        if (colorMode == byClass)
-            color_index = obj.label;
-
-        const unsigned char* color = colors[color_index];
-        cv::Scalar cc(color[0], color[1], color[2]);
-
-        if (colorMode == byIndex)
-            color_index = i;
-
-        cv::rectangle(bgr, obj.rect, cc, 1);
-
-        std::string text = class_names[obj.label] + " " + cv::format("%.2f", obj.prob * 100) + "%";
-
-        int baseLine = 0;
-        cv::Size label_size = cv::getTextSize(text, cv::FONT_HERSHEY_SIMPLEX, 0.5, 1, &baseLine);
-
-        int x = obj.rect.x;
-        int y = obj.rect.y - label_size.height - baseLine;
-        if (y < 0)
-            y = 0;
-        if (x + label_size.width > bgr.cols)
-            x = bgr.cols - label_size.width;
-
-        cv::rectangle(bgr, cv::Rect(cv::Point(x, y), cv::Size(label_size.width, label_size.height + baseLine)), cv::Scalar(255, 255, 255), -1);
-        cv::putText(bgr, text, cv::Point(x, y + label_size.height), cv::FONT_HERSHEY_SIMPLEX, 0.5, cv::Scalar(0, 0, 0));
-
-        draw_mask(bgr, obj.cv_mask, color);
-    }
-}
-
-void Utils::draw_label(cv::Mat& bgr, const cv::Rect2f& rect, std::string label) {
-    int baseLine = 0;
-    cv::Size label_size = cv::getTextSize(label, cv::FONT_HERSHEY_SIMPLEX, 0.5, 1, &baseLine);
-
-    int x = rect.x;
-    int y = rect.y - label_size.height - baseLine;
-    if (y < 0)
-        y = 0;
-    if (x + label_size.width > bgr.cols)
-        x = bgr.cols - label_size.width;
-
-    cv::rectangle(bgr, cv::Rect(cv::Point(x, y), cv::Size(label_size.width, label_size.height + baseLine)), cv::Scalar(255, 255, 255), -1);
-    cv::putText(bgr, label, cv::Point(x, y + label_size.height), cv::FONT_HERSHEY_SIMPLEX, 0.5, cv::Scalar(0, 0, 0));
-}
-
 cv::Mat Utils::applyMask(const cv::Mat& bgr, const cv::Mat& mask) {
     cv::Mat binMask;
     cv::threshold(mask, binMask, 0.5, 255, cv::ThresholdTypes::THRESH_BINARY); // Mask Binarization
@@ -244,40 +102,6 @@ std::vector<cv::Point> Utils::mask2segment(const cv::Mat& mask, int strategy) {
     }
 
     return contour;
-}
-
-void Utils::draw_mask(cv::Mat& bgr, const cv::Mat& mask, const unsigned char* color) {
-    cv::Mat binMask;
-    cv::threshold(mask, binMask, 0.5, 255, cv::ThresholdTypes::THRESH_BINARY); // Mask Binarization
-    for (int y = 0; y < bgr.rows; y++) {
-        uchar* image_ptr = bgr.ptr(y);
-        const float* mask_ptr = binMask.ptr<float>(y);
-        for (int x = 0; x < bgr.cols; x++) {
-            if (mask_ptr[x]) {
-                image_ptr[0] = cv::saturate_cast<uchar>(image_ptr[0] * 0.5 + color[0] * 0.5);
-                image_ptr[1] = cv::saturate_cast<uchar>(image_ptr[1] * 0.5 + color[1] * 0.5);
-                image_ptr[2] = cv::saturate_cast<uchar>(image_ptr[2] * 0.5 + color[2] * 0.5);
-            }
-            image_ptr += 3;
-        }
-    }
-}
-
-void Utils::draw_RotatedRect(cv::Mat& bgr, const cv::RotatedRect& rr, const cv::Scalar& cc, int thickness) {
-    cv::Point2f vertices[4];
-
-    if (padding != 0) {
-        auto new_h = rr.size.height + padding;
-        auto new_w = rr.size.width + padding;
-        cv::RotatedRect padding_rect(rr.center, cv::Size2f(new_w, new_h), rr.angle);
-        padding_rect.points(vertices);
-    }
-
-    else
-        rr.points(vertices);
-
-    for (int i = 0; i < 4; i++)
-        cv::line(bgr, vertices[i], vertices[(i + 1) % 4], cc, thickness);
 }
 
 void Utils::image(const std::filesystem::path& inputPath, const std::filesystem::path& outputFolder) {
@@ -453,7 +277,7 @@ void Utils::video(std::string inputPath) {
                 detector->detect_dynamic(frame, objects, target_size, prob_threshold, agnostic, max_object);
             else
                 detector->detect(frame, objects, target_size, prob_threshold, agnostic, max_object);
-            draw_objects(frame, objects, 0);
+            draw_objects(frame, objects, 0, class_names);
             cv::imshow("Detect", frame);
             if (save) {
                 cv::utils::fs::createDirectory("../frame");
