+++ Yolo/utils.hpp
@@ -8,13 +8,6 @@ enum strategy {
     largestContour = 1          //select largest segment
 };
 
-enum colorMode {
-    byClass = 0,                //color object by class index
-    byIndex = 1                 //color object by object number index 
-};
-
-extern const unsigned char colors[81][3];
-
 extern std::vector<std::string> IMG_FORMATS;
 extern std::vector<std::string> VID_FORMATS;
 
@@ -78,24 +71,7 @@ private:
 
 private:
 
-    /// @brief Draw all the objects at once.
-    /// This function is a combination of draw_mask, draw_label and cv::rectangle
-    /// @param bgr background image to be draw on.
-    /// @param objects object vector contain all the detected object in the image
-    /// @param colorMode determine the color for each object to be draw ( bounding box and feature mask )
-    void draw_objects(cv::Mat& bgr, const std::vector<Object>& objects, int colorMode = byIndex);
-
-    /// @brief draw color mask on the background image.
-    /// @param bgr background image to be draw on.
-    /// @param mask gray scale mask
-    /// @param color color to be draw on the mask
-    void draw_mask(cv::Mat& bgr, const cv::Mat& mask, const unsigned char* color);
-
-    void draw_RotatedRect(cv::Mat& bgr, const cv::RotatedRect& rect, const cv::Scalar& cc, int thickness = 1);
-
     std::vector<cv::Point> mask2segment(const cv::Mat& mask, int strategy = largestContour);
-
-    void draw_label(cv::Mat& bgr, const cv::Rect2f& rect, std::string label);
 
     cv::Mat applyMask(const cv::Mat& bgr, const cv::Mat& mask);
